+++ macshot/Services/HotkeyManager.swift
@@ -20,6 +20,7 @@ class HotkeyManager {
         case scrollCapture = 8
         case openFromClipboard = 9
         case captureLastArea = 10
+        case pinFromClipboard = 11
 
         var keyCodeKey: String {
             switch self {
@@ -33,6 +34,7 @@ class HotkeyManager {
             case .scrollCapture: return "hotkeyScrollCaptureKeyCode"
             case .openFromClipboard: return "hotkeyOpenClipboardKeyCode"
             case .captureLastArea: return "hotkeyCaptureLastAreaKeyCode"
+            case .pinFromClipboard: return "hotkeyPinClipboardKeyCode"
             }
         }
 
@@ -48,6 +50,7 @@ class HotkeyManager {
             case .scrollCapture: return "hotkeyScrollCaptureModifiers"
             case .openFromClipboard: return "hotkeyOpenClipboardModifiers"
             case .captureLastArea: return "hotkeyCaptureLastAreaModifiers"
+            case .pinFromClipboard: return "hotkeyPinClipboardModifiers"
             }
         }
 
@@ -67,6 +70,7 @@ class HotkeyManager {
             case .scrollCapture: return L("Scroll Capture")
             case .openFromClipboard: return L("Open from Clipboard")
             case .captureLastArea: return L("Capture Last Area")
+            case .pinFromClipboard: return L("Pin from Clipboard")
             }
         }
 
@@ -82,12 +86,13 @@ class HotkeyManager {
             case .scrollCapture: return 0
             case .openFromClipboard: return 0  // no default hotkey
             case .captureLastArea: return 0    // no default hotkey
+            case .pinFromClipboard: return 0    // no default hotkey
             }
         }
 
         var defaultModifiers: UInt32 {
             switch self {
-            case .recordScreen, .scrollCapture, .openFromClipboard, .captureLastArea: return 0
+            case .recordScreen, .scrollCapture, .openFromClipboard, .captureLastArea, .pinFromClipboard: return 0
             default: return UInt32(cmdKey | shiftKey)
             }
         }
@@ -126,7 +131,7 @@ class HotkeyManager {
     }
 
     /// Register all hotkeys with their callbacks.
-    func registerAll(captureArea: @escaping () -> Void, captureFullScreen: @escaping () -> Void, recordArea: @escaping () -> Void, recordScreen: @escaping () -> Void, historyOverlay: @escaping () -> Void, captureOCR: @escaping () -> Void, quickCapture: @escaping () -> Void, scrollCapture: @escaping () -> Void, openFromClipboard: @escaping () -> Void, captureLastArea: @escaping () -> Void) {
+    func registerAll(captureArea: @escaping () -> Void, captureFullScreen: @escaping () -> Void, recordArea: @escaping () -> Void, recordScreen: @escaping () -> Void, historyOverlay: @escaping () -> Void, captureOCR: @escaping () -> Void, quickCapture: @escaping () -> Void, scrollCapture: @escaping () -> Void, openFromClipboard: @escaping () -> Void, captureLastArea: @escaping () -> Void, pinFromClipboard: @escaping () -> Void) {
         unregisterAll()
         register(slot: .captureArea, callback: captureArea)
         register(slot: .captureFullScreen, callback: captureFullScreen)
@@ -138,6 +143,7 @@ class HotkeyManager {
         register(slot: .scrollCapture, callback: scrollCapture)
         register(slot: .openFromClipboard, callback: openFromClipboard)
         register(slot: .captureLastArea, callback: captureLastArea)
+        register(slot: .pinFromClipboard, callback: pinFromClipboard)
     }
 
     private func installEventHandler() {